allow ott-collector to discover balancers
