+++ crates/ott-collector/src/main.rs
@@ -1,3 +1,5 @@
+use ott_common::discovery::{start_discovery_task, DnsDiscoveryConfig, DnsMonolithDiscoverer};
+
 #[macro_use]
 extern crate rocket;
 
@@ -9,13 +11,22 @@ fn serve_state() {
     todo!("Serve the current system state")
 }
 
-#[launch]
-fn rocket() -> _ {
-    // TODO: spawn discovery tokio task here
-
-    rocket::build()
+#[rocket::main]
+async fn main() -> Result<(), rocket::Error> {
+    let (discovery_tx, _discovery_rx) = tokio::sync::mpsc::channel(2);
+    let discovery = DnsMonolithDiscoverer::new(DnsDiscoveryConfig {
+        monolith_port: 8081,
+        dns_server: None,
+        query: "balancer".to_string(),
+    });
+    start_discovery_task(discovery, discovery_tx);
+    let _rocket = rocket::build()
         .attach(cors::Cors)
         .mount("/", routes![status, cors::handle_preflight, serve_state])
+        .launch()
+        .await?;
+
+    Ok(())
 }
 
 #[get("/status")]
